Change to config-defined grains
